Addresses #14
Allows users to specifiy a destination profile that is different from the source
profile.

I don't really understand the use case but hey I aim to serve

To specify the destination profile add the destination_profile directive to the config file.  The destination profile can be specified as the profile number or as the profile name

Example:
'''
destination_profile = 2
'''

'''
destination_profile = Ultra-HD
'''